+++ Src/Library/Types.cs
@@ -19,7 +19,6 @@ internal static class Types
     internal static readonly Type EndpointOf3 = typeof(Endpoint<,,>);
     internal static readonly Type EndpointWithMapperOf2 = typeof(EndpointWithMapper<,>);
     internal static readonly Type EndpointWithOutRequestOf2 = typeof(EndpointWithoutRequest<,>);
-    internal static readonly Type FastEventHandlerOf1 = typeof(FastEventHandler<>);
     internal static readonly Type Http = typeof(Http);
     internal static readonly Type HttpAttribute = typeof(HttpAttribute);
     internal static readonly Type ICommandHandler = typeof(ICommandHandler);
@@ -28,6 +27,7 @@ internal static class Types
     internal static readonly Type IEndpointValidator = typeof(IEndpointValidator);
     internal static readonly Type IEnumerable = typeof(IEnumerable);
     internal static readonly Type IEventHandler = typeof(IEventHandler);
+    internal static readonly Type IEventHandlerOf1 = typeof(IEventHandler<>);
     internal static readonly Type IFormFile = typeof(IFormFile);
     internal static readonly Type IHasMapper = typeof(IHasMapper);
     internal static readonly Type IMapper = typeof(IMapper);